# PluginServer/openapi.yaml
openapi: 3.0.3
info:
  title: MicroDock PluginServer API
  version: 2.0.0
  description: |
    MicroDock PluginServer是一个功能完整的插件管理服务器API，提供插件上传、版本管理、备份恢复等核心功能。

    ## 核心特性
    - 插件管理：上传、启用、禁用、删除插件
    - 版本控制：管理插件的多个版本，支持版本回溯
    - 备份管理：用户数据备份和恢复
    - 安全认证：基于JWT的管理员权限控制
    - 文件操作：安全的文件上传下载，支持SHA256校验

    ## API设计原则
    - 仅使用GET和POST方法，确保兼容性
    - URL路径不携带动态参数，提高安全性
    - 统一的JSON响应格式
    - 完整的错误处理机制
  contact:
    name: MicroDock Team
    email: support@microdock.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 开发环境服务器
  - url: https://api.microdock.com
    description: 生产环境服务器
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: 管理员登录
      description: |
        管理员使用用户名和密码登录，获取JWT访问令牌。
        令牌有效期为24小时，需要在需要管理员权限的API请求中携带。
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              adminLogin:
                summary: 管理员登录示例
                value:
                  username: "admin"
                  password: "admin"
      responses:
        '200':
          description: 登录成功，返回JWT令牌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: 获取当前认证状态
      description: 获取当前用户的认证信息和权限状态。可以携带token进行管理员身份验证，也可以不携带token查看公开状态。
      operationId: getCurrentAuthStatus
      security:
        - bearerAuth: []  # 可选认证
      responses:
        '200':
          description: 返回认证状态信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatusResponse'
...
  /api/backups/upload:
    post:
      tags:
        - Backups
      summary: 上传备份文件
      description: |
        用户上传备份文件到服务器。支持程序备份和插件备份两种类型。

        **备份类型说明：**
        - program: MicroDock主程序配置和数据备份
        - plugin: 插件相关数据和配置备份
      operationId: uploadBackup
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BackupUploadRequest'
      responses:
        '200':
          description: 备份文件上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          $ref: '#/components/responses/RequestEntityTooLarge'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/backups/list:
    post:
      tags:
        - Backups
      summary: 获取用户备份列表
      description: 根据用户密钥获取该用户的所有备份文件列表。
      operationId: listUserBackups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupListRequest'
            examples:
              listBackups:
                summary: 获取用户备份列表示例
                value:
                  user_key: "user123abc"
      responses:
        '200':
          description: 备份列表获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/backups/list-all:
    get:
      tags:
        - Backups
      summary: 获取所有备份列表
      description: 获取系统中所有用户的备份文件列表。需要管理员权限。
      operationId: listAllBackups
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 所有备份列表获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/backups/download:
    post:
      tags:
        - Backups
      summary: 下载备份文件
      description: 根据用户密钥和备份ID下载指定的备份文件。
      operationId: downloadBackup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupDownloadRequest'
            examples:
              downloadBackup:
                summary: 下载备份文件示例
                value:
                  user_key: "user123abc"
                  id: 123
      responses:
        '200':
          description: 备份文件下载成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: 文件下载的标头信息
              schema:
                type: string
                example: "attachment; filename=\"backup_program_20231201_143022.zip\""
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/backups/delete:
    post:
      tags:
        - Backups
      summary: 删除备份文件
      description: 删除指定的备份文件。此操作不可逆，请谨慎使用。需要管理员权限。
      operationId: deleteBackup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupDeleteRequest'
            examples:
              deleteBackup:
                summary: 删除备份文件示例
                value:
                  user_key: "user123abc"
                  id: 123
      responses:
        '200':
          description: 备份文件删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  # System endpoints
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT (JSON Web Token) 认证方案

        **获取令牌：** 使用 POST /api/auth/login 获取令牌
        **使用方法：** 在请求头中添加 `Authorization: Bearer <token>`
        **有效期：** 24小时
        **权限说明：** 管理员令牌可以访问所有API，普通用户只能访问公开API

  # Schemas
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 管理员用户名
          example: "admin"
        password:
          type: string
          description: 管理员密码
          example: "admin"
          format: password
...
    BackupListRequest:
      type: object
      required:
        - user_key
      properties:
        user_key:
          type: string
          description: 用户密钥
          example: "user123abc"
          pattern: '^[a-zA-Z0-9._-]+$'
    BackupDownloadRequest:
      type: object
      required:
        - user_key
        - id
      properties:
        user_key:
          type: string
          description: 用户密钥
          example: "user123abc"
          pattern: '^[a-zA-Z0-9._-]+$'
        id:
          type: integer
          description: 备份文件ID
          example: 123
          minimum: 1
    BackupDeleteRequest:
      type: object
      required:
        - user_key
        - id
      properties:
        user_key:
          type: string
          description: 用户密钥
          example: "user123abc"
          pattern: '^[a-zA-Z0-9._-]+$'
        id:
          type: integer
          description: 备份文件ID
          example: 123
          minimum: 1

    # Multipart form data schemas
...
    BackupUploadRequest:
      type: object
      required:
        - file
        - user_key
        - backup_type
      properties:
        file:
          type: string
          format: binary
          description: 备份文件（最大100MB）
        user_key:
          type: string
          description: 用户密钥
          example: "user123abc"
          pattern: '^[a-zA-Z0-9._-]+$'
        backup_type:
          type: string
          description: 备份类型
          enum: [program, plugin]
          example: "program"
        description:
          type: string
          description: 备份描述（可选）
          example: "程序配置备份"

    # Response schemas
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "登录成功"
        token:
          type: string
          description: JWT访问令牌
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: 令牌有效期（秒）
          example: 86400
    AuthStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            is_logged_in:
              type: boolean
              description: 是否已登录
              example: true
            username:
              type: string
              description: 当前登录用户名（仅管理员可见）
              example: "admin"
            is_admin:
              type: boolean
              description: 是否为管理员
              example: true
        message:
          type: string
          example: "认证状态获取成功"
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作成功"

    # Plugin-related response schemas
...
    Backup:
      type: object
      properties:
        id:
          type: integer
          description: 备份ID
          example: 123
        user_key:
          type: string
          description: 用户密钥
          example: "user123abc"
        backup_type:
          type: string
          description: 备份类型
          enum: [program, plugin]
          example: "program"
        file_name:
          type: string
          description: 原始文件名
          example: "backup.zip"
        file_path:
          type: string
          description: 存储路径
          example: "./data/backups/user123abc/program/1701420622_backup.zip"
        file_size:
          type: integer
          description: 文件大小（字节）
          example: 2097152
        file_hash:
          type: string
          description: SHA256文件哈希
          example: "f1e2d3c4b5a6..."
        description:
          type: string
          description: 备份描述
          example: "程序配置备份"
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2023-12-01T14:30:22Z"
    BackupResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "备份上传成功"
        data:
          $ref: '#/components/schemas/Backup'
    BackupListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "备份列表获取成功"
        data:
          type: object
          properties:
            total:
              type: integer
              description: 备份总数
              example: 5
            backups:
              type: array
              items:
                $ref: '#/components/schemas/Backup'

    # System response schemas
...
  responses:
    BadRequest:
      description: 请求参数错误或文件格式不支持
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: 错误详细信息
                example: "文件格式不支持，仅支持.zip格式"
              message:
                type: string
                example: "请求处理失败"
    Unauthorized:
      description: 未授权访问或令牌无效
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: 错误详细信息
                example: "令牌无效或已过期"
              message:
                type: string
                example: "认证失败"
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: 错误详细信息
                example: "需要管理员权限"
              message:
                type: string
                example: "权限不足"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: 错误详细信息
                example: "插件不存在"
              message:
                type: string
                example: "资源未找到"
...
    RequestEntityTooLarge:
      description: 文件过大
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: 错误详细信息
                example: "文件大小超过100MB限制"
              message:
                type: string
                example: "文件过大"
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                description: 错误详细信息
                example: "插件名称格式不正确"
              message:
                type: string
                example: "参数验证失败"